+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+## [0.11.0] - 2026-08-26
+
+- Updated GitHub references from `coltonbh` to `atomsforhumanity`.
+
 ## [0.11.0] - 2026-07-15
 
 ### Fixed
@@ -13,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `decode()` now raises a parser error when output artifacts required by parser specs for the requested calculation type are missing, while preserving any partial data parsed from available artifacts.
 
 ## [0.10.2] - 2026-06-10
+
 - Marked ORCA `parse_natoms` as `required=False` to fix parsing of xTB log files.
 - Fixed gradient regex to account for arbitrary number of stars in log file.
 
@@ -277,6 +282,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Tests for all parsers and the main `parse` function.
 
 [unreleased]: https://github.com/atomsforhumanity/qccodec/compare/0.11.0...HEAD
+[0.11.0]: https://github.com/atomsforhumanity/qccodec/releases/tag/0.11.0
 [0.11.0]: https://github.com/atomsforhumanity/qccodec/releases/tag/0.11.0
 [0.10.2]: https://github.com/atomsforhumanity/qccodec/releases/tag/0.10.2
 [0.10.1]: https://github.com/atomsforhumanity/qccodec/releases/tag/0.10.1